pyproject.toml
[tool.black]
line-length = 110
skip-string-normalization = true
target-version = ['py36', 'py37', 'py38']

[tool.poetry]
authors = ["Michael Statt <michael.statt@modelyst.io>"]
description = "A package designed to perform etl from a postgres database to a neo4j database."
license = "Apache-2.0"
name = "pg4j"
packages = [
  {include = "pg4j", from = "src"},
]
version = "0.1.0"

[tool.poetry.dependencies]
PyYAML = "^5.4.1"
SQLAlchemy = "^1.4.22"
pre-commit = "^2.14.0"
psycopg2-binary = "^2.9.1"
pydantic = "^1.8.2"
python = ">=3.7"
typer = "^0.3.2"
types-PyYAML = "^5.4.6"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
flake8 = "^3.9.2"
mypy = "^0.910"
pytest = "^6.2.4"
GitPython = "^3.1.20"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
